xml data objects and functions to access them

+++ src/realmrelay/data/ItemData.java
@@ -13,7 +13,7 @@ public class ItemData {
 	public boolean usable = false; // element Usable
 	public int bagType = -1; // element BagType
 	public int feedPower = -1; // element feedPower
-	public int rateOfFire = -1; // element RateOfFire
+	public float rateOfFire = -1; // element RateOfFire
 	public int fameBonus = 0; // element FameBonus
 	public int mpCost = -1; // element MpCost
 	public int mpEndCost = -1; // element MpEndCost
